:hammer_and_wrench: 完成JWT签发功能

+++ models/user.go
@@ -3,7 +3,7 @@ package models
 import "github.com/astaxie/beego/orm"
 
 type User struct {
-	Id       int
+	Id       uint64
 	CardId   string
 	NickName string
 	Post     []*Post `orm:"reverse(many)"`
